Introduce Caelus (#1)
This introduces Caelus, an accurate and open-source astronomical ephemeris.

This project is meant to use Astronoby and other tools to provide accurate astronomical data while keeping the code open-source and document the celestial events.

It was inspired by the great [Heavens Above](https://www.heavens-above.com) with the idea to be as transparent as possible about the displayed information:
* What it is semantically
* What it means scientifically
* How it was computed

---

On the technical side, this first change introduces several parts as basis for what will come next:
* `Caelus` module to keep it separate from other sub-projects in the repository
* Use of [Astronoby](https://github.com/rhannequin/astronoby)
* Use of a singleton to load the INPOP19a ephemeris once and have it available anywhere
* Display the distance from the Sun, distance from Earth and velocity of each major Solar System body
* Use i18n
* TailwindCSS for styling
* Have a light/dark mode toggle using Stimulus
* Helper for displaying numerical values
* Code coverage